Removed old Profile Activity card and moved Achievement
